+++ Fractals and Pinballs/Rectangle.cpp
@@ -1,6 +1,6 @@
 #include "Rectangle.h"
 
-Rectangle::Rectangle(std::vector<int> topLeftCorner, int height, int width) {
+Rectangle::Rectangle(std::vector<double> topLeftCorner, double height, double  width) {
 	Rectangle::topLeftCorner = topLeftCorner;
 	Rectangle::height = height;
 	Rectangle::width = width;
